Add "target64".
Add a limited form of 32bit targets that have
pointers and integers that are 64bits in size.

This provides two distinct advantages:
 - It cuts the number of bootstraps in half,
   if bootstrap is defined as buiding just cm3,
   and not the entire system.
   For example the X11 bindings are broken in this scheme.

 - It works around m3front regressions.
   Again, not for the entire system.

It provides "neutral":
 - It continues to work on 32bit hardware
   using 32bit instructions/registers, etc.

It provides not great:
 - Interoperability with native code is reduced.
   In particular, structs get the wrong layout.
   m3core is taught to handle this, with "#if M3_TARGET64'.

 - It doubles some costs.

 - The resulting minimal system is an ABI fork.

 - It is tempting to declare this "the 32bit system",
   but again, it breaks X11 and surely more.

It is not meant to be used with actual 64bit targets,
but should work with them with essentially no change.
Perhaps the passing around of 8 byte structs is
a pessimization. Perhaps.

But again, the main point is to provide a
word-size agnostic .cpp file to build cm3.

The m3front regressions still need to be fixed.